requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
pep8==1.7.1
pytz==2018.9
six==1.12.0
